Relacion de muchos a muchos

+++ Project/src/main/java/mx/LemonTrees/Project/Repository/RanchoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import mx.LemonTrees.Project.Model.Rancho;
 
 @Repository
-public interface RanchoRepository extends CrudRepository<Rancho,Long>{
+public interface RanchoRepository extends CrudRepository<Rancho,Integer>{
